Countdown
* Made the project use the GameMode that was constructed with the C++ project
* GameMode now controls the "Countdown" to start the game.
* Made the pillar Spawner tick, so I can track when the GameMode tells the game starts
* Made the "Bird" pawn tick check when the GameMode says its time to start so it can resolve the Physics.

+++ Source/FlappyBirdClone/ObjectPillar.cpp
@@ -39,7 +39,7 @@ AObjectPillar::AObjectPillar()
 void AObjectPillar::OnConstruction(const FTransform& Transform)
 {
 	float gap = PillarGap / 2;
-	int scale = FMath::FRandRange(1, PillarMaxScale);
+	int scale = FMath::FRandRange(1.f, (float)PillarMaxScale);
 
 	TopPillar->SetWorldScale3D(FVector(1, 1, scale));
 	TopPillar->SetWorldLocation(GetActorLocation() + FVector(0, 0, gap));
